requirements.txt
requests
sqlalchemy
psycopg2-binary
apscheduler
streamlit
vaderSentiment
pandas
plotly
python-dotenv
notebook
